Differentiate handling of warning level enforcement in hhbbc
Summary:
Currently, hhbbc handles readonly error-level enforcement and treats warning-level enforcement and no enforcement the same. We need to treat warning-level enforcement specially in order to not optimize away warnings, but not treat them as terminal as we do with error-level enforcement.

I split up `checkReadonlyOp` into `checkReadonlyOpThrows` and `checkReadonlyOpMaybeThrows` to differentiate these two checks. Since we rename, it is clear that we have revisited each call site.

Reviewed By: oulgen

Differential Revision: D30926093

fbshipit-source-id: bce50400f6e32bc7af404a872d84c08ecc88a61e