Add everything_sdt typechecker option
Summary:
This option annotates every `class` and `function` definition with `<<__SupportDynamicType>>`. This way, with
```
enable_sound_dynamic_type = true
everything_sdt = true
```
we can get a clear picture of what needs to be pessimized in WWW. The similar flag combination over our test cases is useful to expose bugs in the SDT implementation.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D30516607

fbshipit-source-id: ba22643370703597314a7dce0aad121a62bed4b9