Remove 'static requirement in stack_limit library
Summary:
The stack_limit library executes a closure, retrying on a thread with a larger stack size if the closure encounters a stack overflow. Since it spawns threads using std::thread, it must impose the same requirements that std::thread does on its caller: any values borrowed by the closure must never be freed (i.e., they must have the lifetime `'static`). This is a sensible requirement for std::thread, since the thread may never terminate, so it's never safe to free values that are borrowed by it. It's not so sensible for stack_limit, since stack_limit always joins the worker thread before returning. This forces callers to either do redundant work in the job closure (to produce values that otherwise could be shared between re-executions) or lie about the lifetimes of values that are borrowed by the closure (using `unsafe`).

[`crossbeam::scope`](https://docs.rs/crossbeam/0.8.1/crossbeam/fn.scope.html) provides an implementation of "scoped threads" which are guaranteed to have terminated by the end of the invocation of the `scope` function. This allows it to express in its type signature that the closures which run on those threads can borrow non-`'static` values. So if we use crossbeam::scope rather than std::thread in stack_limit, we can remove the `'static` requirement!

This diff also updates a couple callers:

rust_decl_ffi and aast_parser_ffi would previously close over OCaml values (typed as `usize`) representing config structs, and re-convert them from OCaml to Rust at the beginning of each job closure. This is because it was not possible to convert them once ahead of time and just reference them in the job closure (because of the `'static` requirement). We can avoid the redundant conversions now!

Some explicit `'static` lifetimes are also removed from hhbc/compile_ffi.rs, since they're no longer needed.

Reviewed By: shayne-fletcher

Differential Revision: D30943980

fbshipit-source-id: 6e67f31b2358f7eaf68578721321387989323b17