Mark assert_safe unsafe
Summary:
In Rust, it's supposed to be impossible to get UB in code which only invokes safe functions. Providing a safe interface to wrap a value in AssertSafe violates that invariant: one can send a non-Send value to another thread and trigger UB in purely safe code.

Easy fix: if we make the field of AssertSafe private, that makes assert_safe the only way to construct one. Then we can just mark it unsafe.

Reviewed By: shayne-fletcher

Differential Revision: D30937285

fbshipit-source-id: 28da9207c962bcff51c3620fe122706059f71510