Synchronize readonly implementations in typechecker and HackC
Summary:
This diff uses HackC's implementation of local readonly inference instead of the typecheckers in readonly_check.ml in order to synchronize the two. It fixes a few minor bugs that are discrepancies between the two as well, and tests the perf to make sure we are not regressing.

Specifically, this diff has two main parts:
- Add an extra return to rust_parser_errors that tells us whether or not we need to do readonly analysis
- Run readonly analysis during typechecker runs; delete the portion of typechecker code used to infer readonly types and instead depend on the readonly inference in HackC to set explicit readonly expressions everywhere. This just amounts to deleting a lot of typechecker code that is no longer necessary.

Reviewed By: oulgen

Differential Revision: D30847767

fbshipit-source-id: cf04982b67b6272497175d1f6346df0a335dbc02