Add an RAII Suppressor for Class-to-String Conversion Warnings (and use it in a few places.)
Summary:
When class-to-string conversion warnings are enabled, sometimes functionality which cannot be influenced from Hack code itself causes the conversion to happen, such as when serializing an object.  This creates a wrapper that suppresses the warning and makes use of that suppressor in:
- Thrift compact and binary serialization, both for reading and writing.
- Type Structure methods when getting the classname.
- Replacing the previously used boolean for suppression in as.cpp.

I suspect we will find more instances as we move along.

Reviewed By: arnabde03

Differential Revision: D30770530

fbshipit-source-id: b3fe4e4148f283b3a81be4f13e20c896fa16d6b0