Skip SDT parent check on non classes/interfaces
Summary:
Enums are not classes, but they extend the special builtin class `HH\BuiltinEnum`. If this class is marked `<<__SupportDynamicType>>`, you get an error when the enum's parents are checked that the parent is SDT but the enum isn't. However, enums cannot be marked SDT themselves (a different error).

Given the inline comment

> ensure that we implement dynamic if we are a subclass/subinterface of a class/interface

the cleanest solution here is to just not do the parent check when the child class isn't a class or interface.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D30880774

fbshipit-source-id: acc6f4270fef2e613b19af022b20e43b2c8c0b7e