Register capabilities before saving env copy for SDT check
Summary: The environment for SDT checks was being saved before the special locals for coeffects were stored, which raised an error for unbound local during coeffects checking. I considered disabling coeffect checks entirely for the SDT pass, but I'd like to keep them in for the moment as they allow for a stress test of our dependent contexts i.e. `$a::C` when $a is dynamic. If they prove onerous we can turn them off.

Differential Revision: D30881565

fbshipit-source-id: f73472bb9b1701a3d5ef3dcc1b138594dd18ca04