Copy repoAuthType and AttrPersistent from trait property
Summary:
In the case where a class inherits a property from both a base class
and a trait, the property from the trait is considered to have
"won". In almost all cases this is not visible, since we check that
the two properties are identical in almost every way. One case where
it does matter, however, is HHBBC's static analysis. We want to use
the repoAuthType and AttrPersistent bits from the trait property, not
the base class.

This only matters for public properties. It's technically an issue
with both static and non-static properties, but HHBBC doesn't
currently do any analysis for non-static properties, so it's
academic. Fix both cases anyways.

Reviewed By: oulgen

Differential Revision: D30888173

fbshipit-source-id: cf3fa6200954137b46c1f3cced0a23defc390dc1