forbid multiple redefinitions of constants via traits
Summary:
Currently there are three  different semantics for constant inheritance:

1) a child class can shallowly override a parent’s constant
2) a child attempting to override a parent’s constant with a constant from an interface will fatal
3) importing a trait with a similar constant collision makes the trait lose

(Also there is a fatal if a constant is defined in a class and that implements an interface that defines said constant)

This diff adds the --enable-strict-semantics-option that alignes the semantics of 3 with 2.  It only affects the semantics of constants, not that of type constants.

Reviewed By: vassilmladenov

Differential Revision: D30304204

fbshipit-source-id: f104a48254803aaba92a148d7d43dfdfbac5a96d