Refactor heap prefix: auto-create on heap creation
Summary:
To create the illusion of separate heaps, we construct hash keys by combining
the user key string with a unique prefix. Currently this prefix must be created
by users of the API, passed in the `Value` argument to the `NoCache` and
`WithCache` functors.

We only need the prefix value in the `KeyHasher` module, created by `MakeKeyHasher`. We
only create one `KeyHasher` module per `NoCache` (and one `NoCache` per `WithCache`),
so it suffices to simply create our prefix as part of `MakeKeyHasher`, dropping the
requirement from callers and simplifying the public API.

Reviewed By: ljw1004, fahndrich

Differential Revision: D30895739

fbshipit-source-id: 5098856738561c4564cd30c717ee8af95103c0c9