Remove Send requirement in stack_limit library
Summary:
While writing the safety comment in unsafe_utils in the previous diff, I started to wonder whether it was necessary for stack_limit to require job closures to be Send at all. Since the main thread blocks while the job runs, concurrent access won't happen. I think that means that it's safe to close over Send values (and we often do in practice anyway, using `unsafe`).

Instead of encoding the assumption that this pattern is safe in `unsafe` assertions at every use of stack_limit, encode the assumption within the library itself, by removing the `Send` requirement.

Reviewed By: shayne-fletcher

Differential Revision: D30969163

fbshipit-source-id: 5fa5e58ba9ab5882a8ece27f7d29ecdb773bdb6a