Remove UnwindSafe requirement in stack_limit library
Summary:
The `UnwindSafe` requirement is propagated from `std::thread::spawn`, which imposes it to help the user ensure panic safety. That is to say, it prevents the closure from closing over values which are not `UnwindSafe` (meaning that a panic may be raised while the value is in an inconsistent state which violates logical invariants).

It is not necessary for memory safety or to avoid UB. From [the docs](https://doc.rust-lang.org/stable/std/panic/trait.UnwindSafe.html):

> Note, however, that this is not an unsafe trait, so there is not a succinct contract that this trait is providing. Instead it is intended as more of a “speed bump” to alert users of catch_unwind that broken invariants may be witnessed and may need to be accounted for.

In practice, when using the elastic stack library, we are not gaining value from this speed bump. We are just asserting that all closed-over values are panic safe (whether they really are or not). `crossbeam::thread::Scope::spawn` doesn't impose this requirement; with this diff, I'm suggesting that in stack_limit, we shouldn't either.

Reviewed By: shayne-fletcher

Differential Revision: D30969164

fbshipit-source-id: 83272476e19b732cda4a911686a815d495a7400d