hackc_hhas_to_string_cpp_ffi
Summary:
1. Previously we introduced `compile_hhas_from_text` that takes in a source file and returns an `HhasProgram` pointer.
2. This diff adds a function `print_hhas_to_string` that takes this pointer and produces the pretty printed bytecode. The main purpose of this function is bytecode comparison testing.

Add flag `--compile-and-print-hhas` to `hh_single_compile_cpp`. When used, it invokes 1 then 2 to output the pretty printed bytecode. Note that with 1 & 2 combined, this acts the same way as the existing `compile_from_text` today (string -> string), but eventually we want 1 to replace `compile_from_text`, and the flag purely for printing/testing.

Reviewed By: shayne-fletcher

Differential Revision: D30917476

fbshipit-source-id: 2f8232c5e2125d55dac0103b921954c89afdc47f