is_disposable via get_ancestor
Summary:
We have a special case in decls to store whether or not a class implements `IDisposable` or `IAsyncDisposable`, directly, or indirectly (via `require implements`). It's simpler (and as efficient) just to use existing logic (e.g. `has_ancestor`) combined with chasing `requires` types.

Note that the semantics of `requires_ancestor` differs between legacy and shallow decls.

In future we might consider consolidating the various `has_ancestor` functions.

Reviewed By: vsiles

Differential Revision: D30838752

fbshipit-source-id: 2cf46e0633690e93e9c4c898f986a1d6dcd54cd6