Allow multiple filenames in --in-place mode
Summary:
Allow multiple filenames and format all of them in place.

hackfmt picks up .hhconfig from the root directory of the first file given. I didn't bother to re-traverse up to the root for each file in the list here; if filenames in multiple repositories are given, then we'll apply the config from the first repo to all files in the list.

This is arguably not the right behavior, but I'm not sure when it would ever happen to begin with, re-traversing to find .hhconfig and rebuild `Format_env`s would introduce a little perf overhead, and it's more work than I care to do right now.

Reviewed By: DavidSnider

Differential Revision: D30970848

fbshipit-source-id: 4582bfa3d6c392a4a3dc5149a0634619050f10b8