Further improve perf and error messaging
Summary:
This diff looks big, but is actually quite simple: we move all of the methods out of the visitor and into regular functions, now that we don't need the typechecker's readonly check to keep track of any state.

This allows us to improve our error messaging by using parts of the function when there isn't readonly present, i.e. during complex property assignments. Instead of giving a bad error like "this prop is readonly"(which means nothing to a user not familiar with readonly), it can do a minimum amount of readonly analysis necessary to handle assignment.

Reviewed By: oulgen

Differential Revision: D30961623

fbshipit-source-id: 2fd19b311c9e0036a0653e4788b5ec658333cc42