deps_of_file_info returns DepSet.t not list
Summary:
`deps_of_file_info` (prior to this diff) returns a ***list*** of dephashes. If for instance a user was editing a file and they copy+pasted a class definition so it appeared twice, they'd have "class C {}  class C {}" and `deps_of_file_info` would return the exact same dephash twice.

Nothing's wrong with this. All the places where we use the result of `deps_of_file_info`, we first convert it into a DepSet.t, which removes duplicates.

*Almost all*. I had recently written D30783120 (https://github.com/facebook/hhvm/commit/ba072a1061ad5c6587155f18f63ab8a37153de85) which changes `Typing_deps.Files.update_file` to remove old dephashes for a file from `ifiles`, before it adds new deps_of_file_info in. I got telemetry from users about failures, i.e. about trying to remove old dephashes from ifiles but the old dephashes didn't exist.

Here's an example log:
```
[2021-09-14 16:04:51.246] [master] Check kind: will bring hh_server to consistency with code changes, by checking whatever fanout is needed ('Full_check')
[2021-09-14 16:04:51.247] [master] Also reparsing these files with failed naming: GraphTest.php
[2021-09-14 16:04:51.247] [master] Processing changes to 1 file: GraphTest.php
[2021-09-14 16:04:51.247] [master] Processing deferred type decl for 0 file(s), deferred typechecking for 0 file(s)
[2021-09-14 16:04:51.247] [master] Begin Parsing 1 files
[2021-09-14 16:04:51.251] [master] [9k, 3+1ms] GraphTest.php - classes: \GraphTest 0,0
[2021-09-14 16:04:51.251] [master] Parsing 1 files: 0.004691
[2021-09-14 16:04:51.251] [master] Heap size: 152459448
[2021-09-14 16:04:51.251] [master] Begin updating naming tables
[2021-09-14 16:04:51.251] [master] INVARIANT_VIOLATION_BUG [remove_absent_dep] file=GraphTest.php
[2021-09-14 16:04:51.254] [master] UPDATING_DEPS_END (dephash->filename): 0.002886
```

All the INVARIANT_VIOLATION_BUG [remove_absent_dep] I examined from telemetry had the exact same sequence -- they did a check (either full or lazy), they said "Also reparsing these files with failed naming", and then they got the remove_absent_dep bug.

I inspected the code, and formed the hypothesis that the failure was due to a name appearing twice in a file. I tried it out in the IDE and was indeed able to generate this INVARIANT_VIOLATION_BUG.

Differential Revision: D30951618

fbshipit-source-id: dd086b549048b9b3eec5ac247137a10106fbb1a8