Naming_heap.get_filenames_by_hash
Summary: This is part of my mission to replace Typing_deps.Files.get_files with the naming heap. I need to look up a DepSet.t and retrieve the associated filenames. That functionality is already present in Typing_deps.Files.ifiles. This diff adds the exact same functionality, but powered instead by the reverse naming table.

Differential Revision: D30438305

fbshipit-source-id: 7cc8626b1ff863e77b7500d2dd3b3c536ddbf6f1