Needn't use ifiles for 64bit / shmem
Summary:
We now have two ways to turn a set of dephashes into a set of filenames:
* use the reverse naming table, so long as (1) the dephashes are 64bit, i.e. the same hashes that naming-sqlite uses, (2) our naming delta is indexable by dephash, which SharedMem is earlier in the stack but LocalMem used by HackIDE isn't.
* use Typing_deps.Files.ifiles, renamed as Naming_provider.ByHash.ifiles, which is a SECOND reverse naming table that's updated at the same time as the regular reverse naming table, and is stored in master ocaml memory rather than shmem+sqlite.

This diff changes the implementation of Naming_provider.ByHash.get_files: it continues to use Typing_deps.Files.ifiles as always, but also if the conditions are met for using the reverse naming table then it does that also and verifies that the two techniques give identical results.

I've written this diff just as a temporary place along in the diffstack. I don't intend to verify in production.

Differential Revision: D30470723

fbshipit-source-id: 9e01c99098fb1e95b2803bd832dd3836fb0f7bcd