name heap keyed by hash
Summary:
I want the naming table to be used as the dephash->filename map, rather than Typing_deps.Files.ifiles. This will require two things:
1. The reverse naming table in naming-sqlite must support dephash->filename lookups. (This already landed)
2. The in-memory naming heap must support dephash->filename lookups.

Currently the in-memory naming heap uses SymbolName->FileName lookups. This diff therefore changes it to be DepHash->FileName.

There's an irritation. Much of sharedmem depends upon keys being strings. It uses this to prefix the key with the heapname "Naming_funs_<key>", and also uses it to suffix the key with "_old" for the heaps that use oldification. (only the decl heap uses oldification).

**Option1**. I could change sharedmem to support both string keys and 64bit-int-hash keys. For string keys it would allow oldification and would use its internal md5 hashing. For hash keys, it would trust the caller to have made the globally unique in some way, since it couldn't use the prefix.

**Option2**. I could turn a dephash into a string "0xBA2384CF1", and use the existing sharedmem string key library. This means that each key will be hashed once into a dephash, and then a second time by sharedmem/md5 into a sharedmem key.

With this diff I've opted for Option2. I expect it to be just slightly more costly, but we'll see that in the performance analysis. I expect the slightly increased cost to be greatly overshadowed by the lower cost associated with not needing Typing_deps.Files.ifiles.

Differential Revision: D29893531

fbshipit-source-id: da7bd82158700f5c1d7d285a17754e58735e834c