Avoid pessimizing types due to unreachable cond blocks
Summary:
When we do a cond(), the type of the phi-d result is the union of the types of the next() and taken() blocks. However, if either block ends in an unreachable state, we shouldn't include the type of its result in the phi-d type at the merge point.

(Note that it's equivalent to compute the types at the DefLabel based on the predecessors of done_block, but we might as well use the tmps that we already have.)

Reviewed By: ricklavoie

Differential Revision: D31027746

fbshipit-source-id: 4568bf290680ce78c3a54ca05902d073eaafc945