Make shard allocators thread-safe
Summary:
The shard allocator (still named `MapAlloc` in the code) was super unsafe. It used a `RefCell` to protect interior mutability of the control structure, but that only works in a single-threaded context.

Change the `RefCell` to a shared-memory `RwLock` to rule out mysterious crashes or memory corruption in the future.

Also don't pretend like we are ever going to free stuff that was allocated in the shard allocators individually. Freeing stuff explicitly requires keeping track of which allocator allocated the original value. That's just too costly (memory wise).

In the future, we might implement some GC to support reclaiming deleted values.

Reviewed By: mjhostet

Differential Revision: D30764641

fbshipit-source-id: 25cc37e69e8909b03d6ee4ec8ab971512a5b47cd