add JK gate for using old naming table for redecling
Summary: This diff fixes a bug in shallow decl fanout calculation. In particular, we were passing in the post parse naming table into the redecling code. This is problematic because if you delete a file, that naming table will no longer be able to map the filename to the old classes in that file. We fix this by instead passing in the pre-updated forward naming table.

Differential Revision: D31122111

fbshipit-source-id: 8eaf12965ace40762eeda93af95a037ea60a05ac