Update Sites using Trace to Use Folly Logging
Summary:
This switches the logging to use the folly logging library so that the logging settings can be configured at runtime.  By default, all logging currently goes nowhere.

The goal of this was to both switch the call sites but also to evaluate the level of logging that makes sense for each message.  Overwhelming logging is useless logging, so INFO/WARN/ERR are aimed at being useful with the DBG settings being extremely verbose for diagnostic purposes.

Reviewed By: binliu19

Differential Revision: D31072552

fbshipit-source-id: 70db9308ce22fb78f21a37e8b255030b2f982a85