include naming sqlite output in --save-state command
Summary:
This is the first step in getting to ensuring that the dep graph and naming table are built on the same www revision. As the title says, this diff includes the sqlite naming table in the output for the --save-state action.

We see a ~350mb compressed increase in ss size from this change

Differential Revision: D31117247

fbshipit-source-id: c8fd3ab0dd5a979572b3563228de2eabf53d5413