Removing an obsolete <<__RX>> attribute in the typecheck/ test suite
Summary:
File `XHPTest.hhi` included an obsolete `<<__Rx>>` attribute, which this diffs removes.  The test suite did not complain about it because the warning

```
File "XHPTest.hhi", line 4, characters 5-8:
Unrecognized user attribute: `__Rx` starts with __ but is not a standard attribute (Naming[2049])
```

was not reported by `hh_single_type_check` running in `--batch-mode`.

Reviewed By: andrewjkennedy

Differential Revision: D31170878

fbshipit-source-id: e7db550e8256c1b7a34cd9efc1962041bb4abdeb