Remove Callconv
Summary:
# High Level Direction
Now that references are dead, `Callconv` should only appear in a function call position, e.g. the expression in square brackets below:
```
foobar(
  [inout $foo],
  $bar,
  $baz,
);
```

Instead of having a `Callconv` node, let's indicate that function calls have a calling convention associated with each parameter.

# Interesting Changes

* The core change is in `hphp/hack/src/annotated_ast/aast.ml` and literally everything else is fallout.
* `hphp/hack/src/hhbc/hhbc_by_ref/emit_expression.rs`, specifically `emit_args_inout_setters` is probably the scariest bit, but it's a really elegant transformation. Instead of doing a kind of a weird double match we now have a single match

# Motivation

As this diff revealed, the way we handle `inout` in unexpected positions is ... haphazard at best. Often we'll have the following pattern:
```
match node with
| Call (_, _, args, _) -> check_for_callconv_if_needed args
| Callconv _ -> failwith "Unexpected callconv"
...
```
Overly simplified, but this is the gist of it: we'll check for `callconv` when it's relevant, and assume everything falls through to a catch-all case. This _generally_ works, but it's easy to do something like have the catch-all case do nothing, silently, in `hh` but throw in `hhbc`. I found that this happens, for sure, with `isset` (see: T99723886). While *definitely* more verbose, this forces the function call to non function call edge to be explicit and account for this case by either ignoring callconv (which will be obvious) or incorporate it.

Additionally, I've mentioned in a side chat that if we ever decide to add more calling conventions (e.g.: named parameters) I suspect we'd need to do this work anyways, lest we make `Callconv` handling even more fragile.

# Caveats / Concerns

I've left a TODO where I think I've made, either intentionally or unintentionally, HHVM more permissive to `inout` in unexpected positions. I suspect in practice this won't change anything for end user(s) unless they're in the habit of running Hack scripts that don't typecheck. The pattern here would be that we just ignore `inout` when compiling pseudo functions; this is benign given that it just means that the function effectively didn't assign *back* the variable ... which is what is happening because I don't think we have a pseudo function that is *actually* inout.

If I have cycles, I hope to get back to those TODOs and add something akin to:
```
expect_normal_callconv(pk)?;
```
... or ...
```
if not (Ast_defs.param_kind_equal pk Ast_defs.Pnormal)
  Errors.foobar whatever;
```

I'll need to dig into perf, but there are cases where I do something like:
```
List.map ~f:(fun e -> (Ast_defs.Pnormal, e)) el
```
... or ...
```
(* el: (param_kind * Aast.expr) list *)
List.map ~f:snd el
```
I ... kind of doubt it will be an issue in practice, but it could add up?

I've also, unfortunately, changed how some errors w/ inout are reported: if the error used to cite the position on `Callconv`, it now, instead, cites whatever expression was in the `Callconv` node, e.g. before this change you'd see errors with inout arguments of the wrong type look like::
```
foo(inout $x);
//  ^^^^^^^^ error cited here
```
This will now be:
```
foo(inout $x);
//        ^^ error cited here

````

Wilfred and I agree this is a UX regression, but not worth blocking this diff over.

Reviewed By: Wilfred

Differential Revision: D30792968

fbshipit-source-id: ea8cf2a003f9dc6ce038f0361470f35036d974d8