Make `param_kind` non-optional
Summary:
Now that references are dead, `Callconv` should only appear in a function call position, e.g. the expression in square brackets below:
```
foobar(
  [inout $foo],
  $bar,
  $baz,
);
```

Instead of having a `Callconv` node, let's indicate that function calls have a calling convention associated with each parameter.

The first step in doing this is to add `Pnormal` to the `param_kind` type: instead of representing calling convention with an optional param_kind, we explicitly note whether it's `Pnormal` or `Pinout`. We seem to use "normal" to describe this elsewhere (e.g. in `hhbc`).

Reviewed By: Wilfred

Differential Revision: D30495831

fbshipit-source-id: 6c85f636b1582ae7b27b663a360e01bd1d58e228