Add AMIStateROProp to ALocBits
Summary:
In load-elim, for a given instruction with general effects, we take the aliasClass for both stores and kills, and we AND them with the availability ALocBits to clear those bits as they are no longer available because these instructions only *may* load or store.

Previously, we weren't including the ALocBit for `AMIStateROProp` when the alias class was a AMIStateAny. Adding this now.

Reviewed By: ricklavoie

Differential Revision: D31160586

fbshipit-source-id: 13e24ec5202ccc7cacf1dec2abc1fe3d6ef4eb5f