add force_shallow_decl_fanout flag
Summary:
This diff adds a force_shallow_decl_fanout flag to hh_server that enables fanout calculation using shallow decls while maintaining folded decls for actual typechecking.

This is one of the first steps to enabling remote old decls as described here: https://fb.quip.com/PcC0AHHhgFAO

One key thing this diff needs to do is invalidation. In particular, we need to ensure all folded decls are correctly invalidated while doing shallow decl fanout computation. To do this, we take the changed set, look up transitive subclasses, and invalidate all folded decl of those.

You might wonder why we can't simply invalidate the mro_invalidated set. The reason, as summarized by jakebailey is "Shallow fanout tries to be smart, and avoid invalidating MROs which didn't change. If you just add a method, for instance, we shouldn't invalidate the MRO of that class and all its descendants. But we MUST invalidate the folded decl of that class and all its descendants.".

This diff also necessarily fixes a bug in shallow decl fanout calculation. In particular, we were passing in the post parse naming table into the redecling code. This is problematic because if you delete a file, that naming table will no longer be able to map the filename to the old classes in that file. We fix this by instead passing in the pre-updated forward naming table.

Differential Revision: D30912179

fbshipit-source-id: 0bbb6950a3098c824c58d58cbae46a88ea03776b