fix improper tvar bounds update in Typing_escape
Summary:
To remove escaping rigid type variables from lower bounds of a (soft) type variable `tv`, Typing_escape proceeded as such:
1. compute a replacement for the bound `bnd` containing the escaping rigid type variable; we call this replacement `new_bnd`
1. call `subtype new_bnd tv`
3. remove `bnd` from the lower bounds of `tv`

That works in pretty much all the cases except when it happens that `new_bnd <: bnd`; when that is the case, the call to subtype will not modify the lower bounds of `tv` and in step 3 we are just going to kill a legit bound and make the constraints on `tv` weaker. That is unsound.

The exact same problem exists for upper bounds.

The fix in this diff changes the logic to aggregate a list a deletions & additions; performs all deletions first, then all additions (still using subtype calls).

Reviewed By: vsiles

Differential Revision: D31145919

fbshipit-source-id: 7c774d9c7ca08b7766a4a60f445896c4cd7eacec