Fix a SymbolMapTest bug by removing duplication
Summary:
D30711237 (https://github.com/facebook/hhvm/commit/5038c8901d28ed5423de349fea4547b0aa07d794) caused S243962. In theory, this unit test was supposed to catch bugs like that. In practice, this unit test was also bugged - in order to exercise the DB fetching code, we needed to query `m2`, the map which relies on the DB, instead of `m1`, the map which had a bunch of stuff in memory.

This diff fixes that, and refactors the code to use a lambda so it's harder to make the same mistake again.

Reviewed By: bdchik

Differential Revision: D31225583

fbshipit-source-id: 3ccd57181063b3610c30fe573551df9f90f63062