Move inlined builtins into the proper frame
Summary:
Previously, when inlining builtins, we emitted the body of the builtin (the CallBuiltin instruction and associated parameter coercing) after the EndInlining. This was fine, as builtins were not able to observe the contents of their frame. This will no longer be fine with coeffects, as the context local is stored within the frame.

Modify inlining of builtins so that the full body lies within the inlined region. In the process, we eliminate the dumpster fire that was builtin catch traces, and obtain clearer memory effects.

A few notes about some optimizations here:
- Inlined and non-inlined builtins modify their inout arguments in place instead of transiting the stack. This is because a pointer to them is passed to the builtin to modify directly, so there's no hope of store/load elim helping out. LdOutAddrInlined has been introduced to properly track memory effects for this. Adjustments to the frame state have also been made.
- CallBuiltin frequently has both locals and stack slots in its loads set because of this behavior. To avoid excessive spilling, we introduce inout as a separate field of GeneralEffects in order to segregate the stack slots from the local slots.
- I've also separate backtrace locals out into a separate field of GeneralEffects. More work to come on this in future diffs.

Reviewed By: ricklavoie, jano

Differential Revision: D31135478

fbshipit-source-id: df8de8cf519afb6cedf067a2f16c5cfeccbe95d4